requirements.txt
streamlit==1.37.1
ollama>=0.3.0
lancedb>=0.8.0
pyarrow>=15.0.2
pypdf>=4.2.0
requests>=2.32.3
tqdm>=4.66.4
numpy>=1.26.4
pandas>=2.2.2
python-dotenv>=1.0.0
python-docx>=1.2.0
beautifulsoup4>=4.13.0
markdown>=3.9
openpyxl>=3.1.5

# Phase 1: Critical Improvements
tantivy>=0.21.0
xxhash>=3.4.1
faiss-cpu>=1.7.4